Add docs for auto-currying.
