+++ doc/src/language/fn-anon.md
@@ -49,9 +49,13 @@ fn anon_fn_1001(x) { this.data -= x; }
 fn anon_fn_1002() { print this.data; }
 ```
 
+
 WARNING - NOT Closures
 ----------------------
 
 Remember: anonymous functions, though having the same syntax as Rust _closures_, are themselves
 **not** closures.  In particular, they do not capture their running environment.  They are more like
 Rust's function pointers.
+
+They do, however, _capture_ variable _values_ from their execution environment, unless the [`no_capture`]
+feature is turned on.  This is accomplished via [automatic currying][capture].
